requirements.txt
fastmcp>=0.5.0
python-dotenv>=1.0.0
requests>=2.31.0
simplejson>=3.19.1
websocket-client>=1.8.0
pyRofex>=1.2.0
